\name{ex1phxd}
\alias{ex1phxd}
\docType{data}
\title{
Example output data for plotting examples. 
}
\description{
  Phoenix NLME output of simulated 1-compartment PK model.
}
\usage{data(ex1phxd)}
%\format{
%}
\details{
Example output data that can be used to test \code{\link{phxplot}} functions.  
}
%%\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
%%}
%%\references{
%%  ~~ possibly secondary sources and usages ~~
%%}
%\seealso{
%  }
\examples{
data(ex1phxd)
}
\keyword{datasets}
