\name{internal functions}
\alias{appmodest}
\alias{comfunc}
\alias{is.odd}
\alias{roundup}
\alias{getdigits}
\alias{subtractMeans}
\alias{tosent}
\alias{revstr}
\alias{list2mat}
\alias{simmat2vect}
\alias{vshift}
\alias{getmodest}
\alias{phxdata}
\alias{phxvpc.basicplot}
\title{Generic internal functions for phxnlme}
\description{
  These are internal functions for the phxnlme package.
}
\details{
  These are internal phxnlme functions for formatting data for phxnlme-specific plots, and 
  reading values for simulation and creation of VPCs. Not intended for direct use.
}
\value{
  Internal functions.
}
\author{Shuang Liang and Chay Ngee Lim}
\keyword{methods}
